Moving the extend functionality to the base class
